LayoutTests/fast/forms/month-multiple-fields/month-multiple-fields-mouse-events-expected.txt

Multiple fields UI of month input type with mouse events

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Please run this with DRT or WTR.

Test following mouse actions:
Mouse click to focus each of sub-fields
Mouse click on the spin button to update each of sub-fields

==> Focus on the month field.
PASS input.value is "2345-08"
PASS input.value is "2345-09"
PASS input.value is "2345-08"
PASS window.getSelection().rangeCount is 0

==> Focus on the year field.
PASS input.value is "2346-08"
PASS input.value is "2347-08"
PASS input.value is "2346-08"
PASS window.getSelection().rangeCount is 0

==> Click on a disabled field.
PASS input.value is "2346-08"

==> Click on a read-only field.
PASS input.value is "2346-08"

PASS successfullyParsed is true

TEST COMPLETE
